requirements.txt
flask==3.1.0
werkzeug==3.1.3
flask-cors==4.0.0
gunicorn==21.2.0

pillow==11.1.0
pytesseract==0.3.13
PyMuPDF==1.23.8

yt-dlp==2023.12.30
ffmpeg-python==0.2.0
pydub==0.25.1
openai-whisper==20231117

transformers==4.36.2
sentencepiece==0.1.99
numpy==1.26.3
torch==2.5.1